gemini_backend/requirements.txt
flask
flask_cors
google-generativeai